+++ packages/knx-dpts/src/B1/B1.ts
@@ -1,5 +1,4 @@
 import { DataPointAbstract } from '../DataPointAbstract'
-import { KnxReading } from '@repo/knx-common'
 import * as codec from './codec'
 
 export abstract class B1 extends DataPointAbstract<number> {
@@ -15,18 +14,6 @@ export abstract class B1 extends DataPointAbstract<number> {
     /** Send a group write with a numeric value (DPT-specific encoding). */
     public async write(value: number): Promise<void> {
         return this.send(codec.toBuffer(value, Buffer.alloc(this.valueByteLength)))
-    }
-
-    /** Remove a listener previously added with {@link addValueListener}. */
-    public removeValueListener(cb: (reading: KnxReading<number>) => void) {
-        this.valueEvent.removeListener('value-received', cb)
-        this.updateSubscription('value-received')
-    }
-
-    /** Subscribe to incoming group writes and read responses for this address. */
-    public addValueListener(cb: (reading: KnxReading<number>) => void) {
-        this.valueEvent.addListener('value-received', cb)
-        this.updateSubscription('value-received')
     }
 
     public toString(value?: number): string {