+++ packages/knx-dpts/src/B8/B8.ts
@@ -1,5 +1,4 @@
 import { DataPointAbstract } from '../DataPointAbstract'
-import { KnxReading } from '@repo/knx-common'
 import * as codec from './codec'
 
 export abstract class B8 extends DataPointAbstract<number[]> {
@@ -14,16 +13,6 @@ export abstract class B8 extends DataPointAbstract<number[]> {
 
     public async write(bits: number[]): Promise<void> {
         return this.send(codec.toBuffer(bits, Buffer.alloc(this.valueByteLength)))
-    }
-
-    public removeValueListener(cb: (reading: KnxReading<number[]>) => void) {
-        this.valueEvent.removeListener('value-received', cb)
-        this.updateSubscription('value-received')
-    }
-
-    public addValueListener(cb: (reading: KnxReading<number[]>) => void) {
-        this.valueEvent.addListener('value-received', cb)
-        this.updateSubscription('value-received')
     }
 
     public toString(bits?: number[]): string {