+++ packages/knx-dpts/src/Date/Date.ts
@@ -1,4 +1,3 @@
-import { KnxReading } from '@repo/knx-common'
 import { DataPointAbstract } from '../DataPointAbstract'
 import * as codec from './codec'
 
@@ -13,16 +12,6 @@ export abstract class Date extends DataPointAbstract<string> {
 
     public async write(value: string): Promise<void> {
         return this.send(codec.toBuffer(value, Buffer.alloc(this.valueByteLength)))
-    }
-
-    public removeValueListener(cb: (reading: KnxReading<string>) => void) {
-        this.valueEvent.removeListener('value-received', cb)
-        this.updateSubscription('value-received')
-    }
-
-    public addValueListener(cb: (reading: KnxReading<string>) => void) {
-        this.valueEvent.addListener('value-received', cb)
-        this.updateSubscription('value-received')
     }
 
     public toString(value?: string): string {