+++ packages/knx-dpts/src/DateTime/DateTime.ts
@@ -1,4 +1,3 @@
-import { KnxReading } from '@repo/knx-common'
 import { DataPointAbstract } from '../DataPointAbstract'
 import * as codec from './codec'
 import { KnxDateTime } from './types'
@@ -12,16 +11,6 @@ export abstract class DateTime extends DataPointAbstract<KnxDateTime> {
 
     public async write(value: KnxDateTime): Promise<void> {
         return this.send(codec.toBuffer(value, Buffer.alloc(this.valueByteLength)))
-    }
-
-    public removeValueListener(cb: (reading: KnxReading<KnxDateTime>) => void) {
-        this.valueEvent.removeListener('value-received', cb)
-        this.updateSubscription('value-received')
-    }
-
-    public addValueListener(cb: (reading: KnxReading<KnxDateTime>) => void) {
-        this.valueEvent.addListener('value-received', cb)
-        this.updateSubscription('value-received')
     }
 
     public toString(value?: KnxDateTime): string {