+++ packages/knx-dpts/src/F16/F16.ts
@@ -1,4 +1,3 @@
-import { KnxReading } from '@repo/knx-common'
 import { DataPointAbstract } from '../DataPointAbstract'
 import * as codec from './codec'
 
@@ -11,16 +10,6 @@ export abstract class F16 extends DataPointAbstract<number> {
 
     public async write(value: number): Promise<void> {
         return this.send(codec.toBuffer(value, Buffer.alloc(this.valueByteLength)))
-    }
-
-    public removeValueListener(cb: (reading: KnxReading<number>) => void) {
-        this.valueEvent.removeListener('value-received', cb)
-        this.updateSubscription('value-received')
-    }
-
-    public addValueListener(cb: (reading: KnxReading<number>) => void) {
-        this.valueEvent.addListener('value-received', cb)
-        this.updateSubscription('value-received')
     }
 
     public toString(value?: number): string {