+++ packages/knx-dpts/src/V32/V32.ts
@@ -1,5 +1,4 @@
 import { DataPointAbstract } from '../DataPointAbstract'
-import { KnxReading } from '@repo/knx-common'
 import * as codec from './codec'
 
 export abstract class V32 extends DataPointAbstract<number> {
@@ -11,16 +10,6 @@ export abstract class V32 extends DataPointAbstract<number> {
 
     public async write(value: number): Promise<void> {
         return this.send(codec.toBuffer(value, Buffer.alloc(this.valueByteLength)))
-    }
-
-    public removeValueListener(cb: (reading: KnxReading<number>) => void) {
-        this.valueEvent.removeListener('value-received', cb)
-        this.updateSubscription('value-received')
-    }
-
-    public addValueListener(cb: (reading: KnxReading<number>) => void) {
-        this.valueEvent.addListener('value-received', cb)
-        this.updateSubscription('value-received')
     }
 
     public toString(value?: number): string {